adress PR comments
